backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
boto3==1.40.50
botocore==1.40.50
celery==5.3.6
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==5.0.8
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
